anolis: x86/sev: Fix potential sign extension in amd_enc_status_change_finish()
ANBZ: #35414

When npages (int) is left-shifted by PAGE_SHIFT to calculate the size
in bytes, the operation is performed on a signed integer. If npages is
large enough, this can cause:

  1. Signed integer overflow (undefined behavior per C standard)
  2. Sign extension when the result is promoted to unsigned long for
     the enc_dec_hypercall() parameter

For example, In CSV3 guest or SEV-SNP guest, invoke set_memory_encrypted()
or set_memory_encrypted() with param numpages equal or lager than 0x80000
can reproduce the issue. In function amd_enc_status_change_finish(),
left shifting int value 0x80000 12 bits to unsigned long, the actual value
becomes 0xffffffff80000000, which is unexpected.

Cast npages to unsigned long before the shift to ensure:
  - The shift operation uses unsigned arithmetic
  - No sign extension occurs during parameter promotion
  - Correct size calculation for large page counts (>2^19 pages on 32-bit)

Fixes: a1284404dc8e (x86/sev: Make enc_dec_hypercall() accept a size instead of npages)

+++ arch/x86/mm/mem_encrypt_amd.c
@@ -275,7 +275,7 @@ static void amd_enc_status_change_prepare(unsigned long vaddr, int npages, bool 
 static bool amd_enc_status_change_finish(unsigned long vaddr, int npages, bool enc)
 {
 	if (!cc_platform_has(CC_ATTR_HOST_MEM_ENCRYPT))
-		enc_dec_hypercall(vaddr, npages << PAGE_SHIFT, enc);
+		enc_dec_hypercall(vaddr, (unsigned long)npages << PAGE_SHIFT, enc);
 
 	return true;
 }
